Constant  - double , float 데이터타입

+++ 230725/org/opentutorials/javatutorials/variable/ConstantDemo.java
@@ -3,7 +3,7 @@ package org.opentutorials.javatutorials.variable;
 public class ConstantDemo {
 	
 		public static void main(String[] args) {
-			float a = 2.2F;
+			double a = 2.2;
 			System.out.println(a);
 		}
 }
